Use "try: ... finally:" instead of context manager
As support for python 2.7 should be maintained, a different
approach has to be used. `Popen` does not implement a context
manager (i.e. no __enter__ and __exit__ methods). The
alternative is a `try: ... finally: ...` block. Note that I
decided to put `Popen` inside this block. Thus, I had to set
proc to be `None` as the `Popen` constructor itself could also
raise an exception. Whether this approach is actually reasonable
is debatable.